This should fix CI/CD and laptop unittest inconsistencies.
Problem likely due to funny interactions when resolving circular import dependencies.
Solved by rearranging functions and files until circularity problems solved.
