+++ ITR/data/data_providers.py
@@ -11,11 +11,11 @@ from pint import Quantity
 from ITR.data.osc_units import ureg
 
 from ITR.interfaces import ICompanyData, EScope, IHistoricData, IProductionRealization, IHistoricEmissionsScopes, \
-    IHistoricEIScopes, ICompanyEIProjection, ICompanyEIProjectionsScopes, ICompanyEIProjections, TemperatureScoreControls
+    IHistoricEIScopes, ICompanyEIProjection, ICompanyEIProjectionsScopes, ICompanyEIProjections
 
 from ITR.interfaces import ICompanyData
 
-from ITR.configs import TabsConfig, ColumnsConfig, VariablesConfig, TemperatureScoreConfig 
+from ITR.configs import TabsConfig, ColumnsConfig, VariablesConfig, TemperatureScoreControls, TemperatureScoreConfig 
 
 class CompanyDataProvider(ABC):
     """
