+++ ITR/interfaces.py
@@ -8,509 +8,14 @@ from enum import Enum
 from typing import Optional, Dict, List, Literal, Union
 from typing import TYPE_CHECKING, Callable
 from pydantic import BaseModel, parse_obj_as, validator, root_validator
-from dataclasses import dataclass
 
 import ITR
 from ITR.logger import logger
-from ITR.data.osc_units import ureg, Q_, M_
+from ITR.data.osc_units import ureg, Q_, M_, BenchmarkMetric, BenchmarkQuantity, ProductionMetric, ProductionQuantity, EmissionsMetric, EmissionsQuantity, EI_Metric, EI_Quantity, quantity
+from ITR.configs import ProjectionControls
 
 import pint
 from pint.errors import DimensionalityError
-
-@dataclass
-class ProjectionControls:
-    LOWER_PERCENTILE: float = 0.1
-    UPPER_PERCENTILE: float = 0.9
-
-    LOWER_DELTA: float = -0.10
-    UPPER_DELTA: float = +0.03
-
-    # FIXME: Should agree with TemperatureScoreConfig.CONTROLS_CONFIG
-    BASE_YEAR: int = 2019
-    TARGET_YEAR: int = 2050
-    TREND_CALC_METHOD: Callable[[pd.DataFrame], pd.DataFrame] = staticmethod(pd.DataFrame.median)
-
-
-class PintModel(BaseModel):
-    class Config:
-        arbitrary_types_allowed = True
-
-# List of all the production units we know
-_production_units = [ "Wh", "pkm", "tkm", "boe", "t Aluminum", "t Cement", "t Copper", "t Paper", "t Steel", "USD", "m**2" ]
-_ei_units = [f"t CO2/({pu})" if ' ' in pu else f"t CO2/{pu}" for pu in _production_units]
-
-class ProductionMetric(str):
-    """
-    Valid production metrics accepted by ITR tool
-    """
-
-    @classmethod
-    def __get_validators__(cls):
-        # one or more validators may be yielded which will be called in the
-        # order to validate the input, each validator will receive as an input
-        # the value returned from the previous validator
-        yield cls.validate
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        # __modify_schema__ should mutate the dict it receives in place,
-        # the returned value will be ignored
-        field_schema.update(
-            examples=_production_units,
-        )
-
-    @classmethod
-    def validate(cls, units):
-        if not isinstance(units, str):
-            raise TypeError('string required')
-        qty = ureg(units)
-        for pu in _production_units:
-            if qty.is_compatible_with(pu):
-                return cls(units)
-        raise ValueError(f"{v} not relateable to {_production_units}")
-
-    def __repr__(self):
-        return f"ProductionMetric({super().__repr__()})"
-
-class EmissionsMetric(str):
-    """
-    Valid production metrics accepted by ITR tool
-    """
-
-    @classmethod
-    def __get_validators__(cls):
-        # one or more validators may be yielded which will be called in the
-        # order to validate the input, each validator will receive as an input
-        # the value returned from the previous validator
-        yield cls.validate
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        # __modify_schema__ should mutate the dict it receives in place,
-        # the returned value will be ignored
-        field_schema.update(
-            examples=['g CO2', 'kg CO2', 't CO2', 'Mt CO2'],
-        )
-
-    @classmethod
-    def validate(cls, units):
-        if not isinstance(units, str):
-            raise TypeError('string required')
-        qty = ureg(units)
-        if qty.is_compatible_with('t CO2'):
-            return cls(units)
-        raise ValueError(f"{units} not relateable to 't CO2'")
-
-    def __repr__(self):
-        return f'ProductionMetric({super().__repr__()})'
-
-class EI_Metric(str):
-    """
-    Valid production metrics accepted by ITR tool
-    """
-
-    @classmethod
-    def __get_validators__(cls):
-        # one or more validators may be yielded which will be called in the
-        # order to validate the input, each validator will receive as an input
-        # the value returned from the previous validator
-        yield cls.validate
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        # __modify_schema__ should mutate the dict it receives in place,
-        # the returned value will be ignored
-        field_schema.update(
-            examples=['g CO2/pkm', 'kg CO2/tkm', 't CO2/(t Steel)', 'Mt CO2/TWh'],
-        )
-
-    @classmethod
-    def validate(cls, units):
-        if not isinstance(units, str):
-            raise TypeError('string required')
-        qty = ureg(units)
-        for ei_u in _ei_units:
-            if qty.is_compatible_with(ei_u):
-                return cls(units)
-        raise ValueError(f"{units} not relateable to {_ei_units}")
-
-    def __repr__(self):
-        return f'ProductionMetric({super().__repr__()})'
-
-class BenchmarkMetric(str):
-    """
-    Valid benchmark metrics accepted by ITR tool
-    """
-
-    @classmethod
-    def __get_validators__(cls):
-        # one or more validators may be yielded which will be called in the
-        # order to validate the input, each validator will receive as an input
-        # the value returned from the previous validator
-        yield cls.validate
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        # __modify_schema__ should mutate the dict it receives in place,
-        # the returned value will be ignored
-        field_schema.update(
-            examples=['dimensionless', 'g CO2/pkm', 'kg CO2/tkm', 't CO2/(t Steel)', 'Mt CO2/TWh'],
-        )
-
-    @classmethod
-    def validate(cls, units):
-        if not isinstance(units, str):
-            raise TypeError('string required')
-        if units=='dimensionless':
-            return cls(units)
-        qty = ureg(units)
-        for ei_u in _ei_units:
-            if qty.is_compatible_with(ei_u):
-                return cls(units)
-        raise ValueError(f"{units} not relateable to 'dimensionless' or {_ei_units}")
-
-    def __repr__(self):
-        return f'ProductionMetric({super().__repr__()})'
-
-# Borrowed from https://github.com/hgrecco/pint/issues/1166
-registry = ureg
-
-schema_extra = dict(definitions=[
-    dict(
-        Quantity=dict(type="string"),
-        # We don't need EmissionsQuantity separately because Quantity does all we nmeed
-        ProductionQuantity=dict(type="List[str]"),
-        EI_Quantity=dict(type="List[str]"),
-    )
-])
-
-
-def quantity(dimensionality: str) -> type:
-    """A method for making a pydantic compliant Pint quantity field type."""
-
-    try:
-        registry.get_dimensionality(dimensionality)
-    except KeyError:
-        raise ValueError(f"{dimensionality} is not a valid dimensionality in pint!")
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value):
-        if isinstance(value, str):
-            try:
-                q = Q_(value)
-            except ValueError:
-                raise ValueError(f"cannot convert '{quantity}' to quantity")
-            quantity = q
-        elif isinstance(value, pint.Quantity):
-            quantity = value
-        else:
-            raise TypeError (f"quantity takes either a Q_ value or a string fully expressing a quantified value; got {value}")
-        if quantity.is_compatible_with(dimensionality):
-            return quantity
-        assert quantity.check(cls.dimensionality), f"Dimensionality of {quantity} incompatible with {cls.dimensionality}"
-        return quantity
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        field_schema.update(
-            {"$ref": "#/definitions/Quantity"}
-        )
-    
-    return type(
-        "Quantity",
-        (pint.Quantity,),
-        dict(
-            __get_validators__=__get_validators__,
-            __modify_schema__=__modify_schema__,
-            dimensionality=dimensionality,
-            validate=validate,
-        ),
-    )
-# end of borrowing
-
-
-class EmissionsQuantity(pint.Quantity):
-    """A method for making a pydantic compliant Pint emissions quantity."""
-
-    def __new__(cls, value, units=None):
-        # Re-used the instance we are passed.  Do we need to copy?
-        return value
-
-    @classmethod
-    def __get_validators__(cls):
-        # one or more validators may be yielded which will be called in the
-        # order to validate the input, each validator will receive as an input
-        # the value returned from the previous validator
-        yield cls.validate
-
-
-    @classmethod
-    def validate(cls, quantity):
-        if quantity is None:
-            breakpoint()
-        if isinstance(quantity, str):
-            v, u = quantity.split(' ', 1)
-            try:
-                q = Q_(float(v), u)
-            except ValueError:
-                raise ValueError(f"cannot convert '{quantity}' to quantity")
-            quantity = q
-        if not isinstance(quantity, pint.Quantity):
-            raise TypeError(f"pint.Quantity required ({quantity}, type = {type(quantity)})")
-        if quantity.is_compatible_with('t CO2'):
-            return quantity
-        raise DimensionalityError (quantity, 't CO2', dim1='', dim2='', extra_msg=f"Dimensionality must be compatible with 't CO2'")
-
-    distance: quantity("[length]")
-    speed: quantity("[length]/[time]")
-
-    class Config:
-        validate_assignment = True
-        schema_extra = schema_extra
-        json_encoders = {
-            pint.Quantity: str,
-        }
-# end of borrowing
-
-
-def emissions_quantity(dimensionality: str) -> type:
-    """A method for making a pydantic compliant Pint quantity field type."""
-
-    def __new__(cls, value, units=None):
-        # Re-used the instance we are passed.  Do we need to copy?
-        return value
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value):
-        if isinstance(value, dict):
-            quantity = Q_(value['units'])
-        else:
-            quantity = Q_(value)
-        if quantity.is_compatible_with(cls.dimensionality):
-            return quantity
-        assert quantity.check(cls.dimensionality), f"Dimensionality must be {cls.dimensionality} {breakpoint()}"
-        return quantity
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        field_schema.update(
-            {"$ref": "#/definitions/EmissionsQuantity"}
-        )
-    
-    return type(
-        "Quantity",
-        (pint.Quantity,),
-        dict(
-            __get_validators__=__get_validators__,
-            __modify_schema__=__modify_schema__,
-            dimensionality=dimensionality,
-            validate=validate,
-        ),
-    )
-
-    @classmethod
-    def validate(cls, quantity):
-        if quantity is None:
-            breakpoint()
-        if isinstance(quantity, str):
-            v, u = quantity.split(' ', 1)
-            try:
-                q = Q_(float(v), u)
-            except ValueError:
-                raise ValueError(f"cannot convert '{quantity}' to quantity")
-            quantity = q
-        if not isinstance(quantity, pint.Quantity):
-            raise TypeError('pint.Quantity required')
-        for pu in _production_units:
-            if quantity.is_compatible_with(pu):
-                return quantity
-        raise DimensionalityError (quantity, str(_production_units), dim1='', dim2='', extra_msg=f"Dimensionality must be compatible with [{_production_units}]")
-
-    @validator('dims_list')
-    def units_must_be_registered(cls, v):
-        for d in v:
-            try:
-                registry.get_dimensionality(d)
-            except KeyError:
-                raise ValueError(f"{d} is not a valid dimensionality in pint!")
-        return v
-
-    class Config:
-        validate_assignment = True
-        schema_extra = schema_extra
-        json_encoders = {
-            pint.Quantity: str,
-        }
-
-def production_quantity(dims_list: List[str]) -> type:
-    """A method for making a pydantic compliant Pint production quantity."""
-
-class EI_Quantity(str):
-    """A method for making a pydantic compliant Pint Emissions Intensity quantity."""
-
-    def __new__(cls, value, units=None):
-        # Re-used the instance we are passed.  Do we need to copy?
-        return value
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value):
-        if isinstance(value, dict):
-            quantity = Q_(value['units'])
-        else:
-            quantity = Q_(value)
-        for dimensionality in ProductionQuantity(dims_list=dims_list).dims_list:
-            if quantity.check(dimensionality):
-                return quantity
-        raise DimensionalityError(value.units, f"in [{ProductionQuantity(dims_list).dims_list}]")
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        field_schema.update(
-            {"$ref": "#/definitions/ProductionQuantity"}
-        )
-    
-    return type(
-        "ProductionQuantity",
-        (ProductionQuantity,),
-        dict(
-            __get_validators__=__get_validators__,
-            __modify_schema__=__modify_schema__,
-            dims_list=f"List[str] = {dims_list}",
-            validate=validate,
-        ),
-    )
-
-    @classmethod
-    def validate(cls, quantity):
-        if quantity is None:
-            breakpoint()
-        if isinstance(quantity, str):
-            v, u = quantity.split(' ', 1)
-            try:
-                q = Q_(float(v), u)
-            except ValueError:
-                raise ValueError(f"cannot convert '{quantity}' to quantity")
-            quantity = q
-        if not isinstance(quantity, pint.Quantity):
-            raise TypeError('pint.Quantity required')
-        for ei_u in _ei_units:
-            if quantity.is_compatible_with(ei_u):
-                return quantity
-        raise DimensionalityError (quantity, str(_ei_units), dim1='', dim2='', extra_msg=f"Dimensionality must be compatible with [{_ei_units}]")
-
-class EI_Quantity(BaseModel):
-
-    dims_list: List[str]
-
-    @validator('dims_list')
-    def units_must_be_registered(cls, v):
-        for d in v:
-            try:
-                registry.get_dimensionality(d)
-            except KeyError:
-                raise ValueError(f"{d} is not a valid dimensionality in pint!")
-        return v
-
-    class Config:
-        validate_assignment = True
-        schema_extra = schema_extra
-        json_encoders = {
-            pint.Quantity: str,
-        }
-
-def ei_quantity(dims_list: List[str]) -> type:
-    """A method for making a pydantic compliant Pint Emissions Intensity (EI) quantity."""
-
-class BenchmarkQuantity(str):
-    """A method for making a pydantic compliant Pint Benchmark quantity (which includes dimensionless production growth)."""
-
-    def __new__(cls, value, units=None):
-        # Re-used the instance we are passed.  Do we need to copy?
-        return value
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value):
-        if isinstance(value, dict):
-            quantity = Q_(value['units'])
-        else:
-            quantity = Q_(value)
-        for dimensionality in EI_Quantity(dims_list=dims_list).dims_list:
-            if quantity.check(dimensionality):
-                return quantity
-        raise DimensionalityError(value.units, f"in [{EI_Quantity(dims_list).dims_list}]")
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        field_schema.update(
-            {"$ref": "#/definitions/EI_Quantity"}
-        )
-    
-    return type(
-        "EI_Quantity",
-        (EI_Quantity,),
-        dict(
-            __get_validators__=__get_validators__,
-            __modify_schema__=__modify_schema__,
-            dims_list=f"List[str] = {dims_list}",
-            validate=validate,
-        ),
-    )
-
-    @classmethod
-    def validate(cls, quantity):
-        if isinstance(quantity, str):
-            v, u = quantity.split(' ', 1)
-            try:
-                q = Q_(float(v), u)
-            except ValueError:
-                raise ValueError(f"cannot convert '{quantity}' to quantity")
-            quantity = q
-        if not isinstance(quantity, pint.Quantity):
-            raise TypeError('pint.Quantity required')
-        if str(quantity.u) == 'dimensionless':
-            return quantity
-        for ei_u in _ei_units:
-            if quantity.is_compatible_with(ei_u):
-                return quantity
-        raise DimensionalityError (quantity, str(_ei_units), dim1='', dim2='', extra_msg=f"Dimensionality must be 'dimensionless' or compatible with [{_ei_units}]")
-
-
-class OSC_Metric(BaseModel):
-    units: str 
-    @validator('units')
-    def units_must_be_OSC(cls, v):
-        if v == 'dimensionless':
-            return v
-        try:
-            if ProductionMetric.unit_must_be_production(v):
-                return v
-        except ValueError:
-            try:
-                if EmissionsMetric.units_must_be_tCO2(v):
-                    return v
-            except ValueError:
-                try:
-                    if IntensityMetric.units_must_be_EI(v):
-                        return v
-                except ValueError:
-                    raise ValueError(f"cannot understand {v} as OSC_Metric")
-
 
 class SortableEnum(Enum):
     def __str__(self):
@@ -1047,22 +552,3 @@ class ICompanyAggregates(ICompanyData):
     # projected_targets: Optional[ICompanyEIProjectionsScopes]
     # projected_intensities: Optional[ICompanyEIProjectionsScopes]
 
-class TemperatureScoreControls(BaseModel):
-    base_year: int
-    target_end_year: int
-    projection_start_year: int
-    projection_end_year: int
-    tcre: quantity('delta_degC')
-    carbon_conversion: emissions_quantity('t CO2')
-    scenario_target_temperature: quantity('delta_degC')
-
-    def __getitem__(self, item):
-        return getattr(self, item)
-
-    @property
-    def tcre_multiplier(self) -> quantity('delta_degC/(t CO2)'):
-        return self.tcre / self.carbon_conversion
-
-# FIXME: Can somebody help sort out the circularities we have?
-IEIRealization.update_forward_refs()
-TemperatureScoreControls.update_forward_refs()
