+++ ITR/temperature_score.py
@@ -35,7 +35,6 @@ class TemperatureScore(PortfolioAggregation):
                  grouping: Optional[List] = None, config: Type[TemperatureScoreConfig] = TemperatureScoreConfig):
         super().__init__(config)
         self.c: Type[TemperatureScoreConfig] = config
-        assert self.c.CONTROLS_CONFIG.tcre == Q_(2.2, 'delta_degC')
         self.fallback_score = fallback_score
 
         self.time_frames = time_frames
@@ -72,12 +71,6 @@ class TemperatureScore(PortfolioAggregation):
                 (scorable_row[self.c.COLS.BENCHMARK_GLOBAL_BUDGET] * (trajectory_overshoot_ratio - 1.0) *
                     self.c.CONTROLS_CONFIG.tcre_multiplier)
             score = trajectory_temperature_score
-            print(f"""trajectory_overshoot_ratio = {scorable_row[self.c.COLS.CUMULATIVE_TRAJECTORY]} / {scorable_row[
-                self.c.COLS.CUMULATIVE_BUDGET]} = {trajectory_overshoot_ratio}""")
-            print(f"""trajectory_temperature_score = {scorable_row[self.c.COLS.BENCHMARK_TEMP]} + \
-                ({scorable_row[self.c.COLS.BENCHMARK_GLOBAL_BUDGET]} * ({trajectory_overshoot_ratio} - 1.0) *
-                    {self.c.CONTROLS_CONFIG.tcre_multiplier}) = {trajectory_temperature_score}""")
-            print(f"score = {trajectory_temperature_score}")
             return score, trajectory_temperature_score, trajectory_overshoot_ratio, \
                 target_temperature_score, target_overshoot_ratio, EScoreResultType.TRAJECTORY_ONLY
